pyproject.toml
[project]
name = "gsync"
version = "0.1.0"
description = "A Google Tasks <-> Markdown sync tool."
dependencies = [
    "google-api-python-client>=2.118.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "PyYAML>=6.0.1",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.2.2",
    "pyright>=1.1.350",
    "google-api-python-client-stubs>=1.24.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/gsync"]

[tool.pyright]
typeCheckingMode = "strict"
